:fire:Fix: PostPage 모달 수정, 글 작성시 관련된 카테고리로 네비게이트. by@Tummimi
